requirements.txt
fastapi
uvicorn[standard]
pydantic
litellm
python-dotenv
requests